+++ AprsClientExample/AprsClientExample.csproj
@@ -12,6 +12,6 @@
     <LangVersion>Latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Boerman.AprsClient\Boerman.AprsClient.csproj" />
+    <ProjectReference Include="..\Skyhop.AprsClient\Skyhop.AprsClient.csproj" />
   </ItemGroup>
 </Project>
