+++ AprsClientExample/Program.cs
@@ -1,9 +1,9 @@
-﻿using System;
+﻿using Skyhop.AprsClient;
+using Skyhop.AprsClient.Models;
+using System;
 using System.Globalization;
 using System.IO;
 using System.Threading.Tasks;
-using Boerman.AprsClient;
-using Boerman.AprsClient.Models;
 
 namespace AprsClientExample
 {
