--- Boerman.AprsClient.sln
@@ -1,65 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.16
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Boerman.AprsClient", "Boerman.AprsClient\Boerman.AprsClient.csproj", "{239C1EC4-26CF-4393-8609-AC095FB04D76}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AprsClientExample", "AprsClientExample\AprsClientExample.csproj", "{A35CDB81-9579-4D70-8B7F-6412F2025539}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Boerman.AprsClient.Tests", "Boerman.AprsClient.Tests\Boerman.AprsClient.Tests.csproj", "{CE331010-10E7-401B-A7CA-39407E947059}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Debug|x64.Build.0 = Debug|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Debug|x86.Build.0 = Debug|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Release|Any CPU.Build.0 = Release|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Release|x64.ActiveCfg = Release|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Release|x64.Build.0 = Release|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Release|x86.ActiveCfg = Release|Any CPU
-		{239C1EC4-26CF-4393-8609-AC095FB04D76}.Release|x86.Build.0 = Release|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Debug|x64.Build.0 = Debug|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Debug|x86.Build.0 = Debug|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Release|x64.ActiveCfg = Release|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Release|x64.Build.0 = Release|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Release|x86.ActiveCfg = Release|Any CPU
-		{A35CDB81-9579-4D70-8B7F-6412F2025539}.Release|x86.Build.0 = Release|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Debug|x64.Build.0 = Debug|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Debug|x86.Build.0 = Debug|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Release|x64.ActiveCfg = Release|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Release|x64.Build.0 = Release|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Release|x86.ActiveCfg = Release|Any CPU
-		{CE331010-10E7-401B-A7CA-39407E947059}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4687D67A-40E9-4C13-A819-5A1636493782}
-	EndGlobalSection
-EndGlobal
