--- Boerman.AprsClient/Boerman.AprsClient.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <Authors>Corstian Boerman</Authors>
-    <Company>The Boerman Company</Company>
-    <Product />
-    <Copyright>Corstian Boerman</Copyright>
-    <PackageProjectUrl>https://github.com/Boerman/Boerman.AprsClient</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/Boerman/Boerman.AprsClient</RepositoryUrl>
-    <PackageReleaseNotes>Updated dependencies for additional stability.</PackageReleaseNotes>
-    <PackageIconUrl>https://avatars2.githubusercontent.com/u/25320889</PackageIconUrl>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.2.7</Version>
-    <Description>The APRS Client library is a library which will get you up and running in a snap when your goal is to read APRS feeds.</Description>
-    <PackageTags>aprs client tcp boerman library glider glidernet ogn openglidernet open net performant easy to use</PackageTags>
-    
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.1.1" />
-    <PackageReference Include="Boerman.Networking" Version="1.1.2" />
-    <PackageReference Include="Boerman.Core" Version="2.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Remove="Extensions.cs" />
-  </ItemGroup>
-</Project>
