--- Boerman.AprsClient/Enums/MicEMessageType.cs
@@ -1,22 +1,0 @@
-﻿namespace Boerman.AprsClient.Enums
-{
-    public enum MicEMessageType
-    {
-        OffDuty,
-        EnRoute,
-        InService,
-        Returning,
-        Committed,
-        Special,
-        Priority,
-        Emergency,
-        Custom0,
-        Custom1,
-        Custom2,
-        Custom3,
-        Custom4,
-        Custom5,
-        Custom6,
-        Unknown
-    }
-}
