--- Boerman.AprsClient/Enums/Symbol.cs
@@ -1,7 +1,0 @@
-﻿namespace Boerman.AprsClient.Enums
-{
-    public enum Symbol
-    {
-        Aircraft
-    }
-}
