--- Boerman.AprsClient/Enums/SymbolTable.cs
@@ -1,8 +1,0 @@
-﻿namespace Boerman.AprsClient.Enums
-{
-    public enum SymbolTable
-    {
-        Primary,
-        Secondary
-    }
-}
