--- Boerman.AprsClient/Models/AprsDataReceivedEventArgs.cs
@@ -1,12 +1,0 @@
-﻿namespace Boerman.AprsClient.Models
-{
-    public class AprsDataReceivedEventArgs
-    {
-        public AprsDataReceivedEventArgs(string data)
-        {
-            Data = data;
-        }
-
-        public string Data { get; set; }
-    }
-}
