--- Boerman.AprsClient/Models/OgnAprsMessage.cs
@@ -1,8 +1,0 @@
-﻿namespace Boerman.AprsClient.Models
-{
-    public class OgnAprsMessage : AprsMessage
-    {
-        public double Climbrate { get; internal set; }
-        public double Turnrate { get; internal set; }
-    }
-}
