--- Boerman.AprsClient/Models/PacketInfoEventArgs.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Boerman.AprsClient.Models
-{
-    public class PacketReceivedEventArgs : EventArgs
-    {
-        public PacketReceivedEventArgs(AprsMessage pi)
-        {
-            AprsMessage = pi;
-        }
-
-        public AprsMessage AprsMessage { get; private set; }
-    }
-}
