+++ Skyhop.AprsClient.Tests/FilterBuilderTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Skyhop.AprsClient;
 
 namespace Boerman.AprsClient.Tests
 {
