+++ Skyhop.AprsClient/Enums/MicEMessageType.cs
@@ -1,0 +1,22 @@
+﻿namespace Skyhop.AprsClient.Enums
+{
+    public enum MicEMessageType
+    {
+        OffDuty,
+        EnRoute,
+        InService,
+        Returning,
+        Committed,
+        Special,
+        Priority,
+        Emergency,
+        Custom0,
+        Custom1,
+        Custom2,
+        Custom3,
+        Custom4,
+        Custom5,
+        Custom6,
+        Unknown
+    }
+}
