+++ Skyhop.AprsClient/Enums/Symbol.cs
@@ -1,0 +1,7 @@
+﻿namespace Skyhop.AprsClient.Enums
+{
+    public enum Symbol
+    {
+        Aircraft
+    }
+}
