+++ Skyhop.AprsClient/Enums/SymbolTable.cs
@@ -1,0 +1,8 @@
+﻿namespace Skyhop.AprsClient.Enums
+{
+    public enum SymbolTable
+    {
+        Primary,
+        Secondary
+    }
+}
