+++ Skyhop.AprsClient/Models/AprsDataReceivedEventArgs.cs
@@ -1,0 +1,12 @@
+﻿namespace Skyhop.AprsClient.Models
+{
+    public class AprsDataReceivedEventArgs
+    {
+        public AprsDataReceivedEventArgs(string data)
+        {
+            Data = data;
+        }
+
+        public string Data { get; set; }
+    }
+}
