+++ Skyhop.AprsClient/Models/OgnAprsMessage.cs
@@ -1,0 +1,8 @@
+﻿namespace Skyhop.AprsClient.Models
+{
+    public class OgnAprsMessage : AprsMessage
+    {
+        public double Climbrate { get; internal set; }
+        public double Turnrate { get; internal set; }
+    }
+}
